+++ package.json
@@ -59,22 +59,22 @@
 		"@babel/runtime": "^7.14.0",
 		"@commitlint/cli": "^17.0.3",
 		"@commitlint/config-conventional": "^17.0.3",
-		"@storybook/addon-actions": "^6.5.9",
-		"@storybook/addon-docs": "^6.5.9",
-		"@storybook/addon-essentials": "^6.5.9",
-		"@storybook/addon-links": "^6.5.9",
-		"@storybook/react": "^6.5.9",
+		"@storybook/addon-actions": "^6.5.12",
+		"@storybook/addon-docs": "^6.5.12",
+		"@storybook/addon-essentials": "^6.5.12",
+		"@storybook/addon-links": "^6.5.12",
+		"@storybook/react": "^6.5.12",
 		"@testing-library/jest-dom": "^5.16.1",
 		"@testing-library/react": "^12.1.5",
 		"@testing-library/react-hooks": "^8.0.1",
-		"@testing-library/user-event": "^14.2.1",
-		"@types/jest": "^26.0.24",
+		"@testing-library/user-event": "^14.4.3",
+		"@types/jest": "^29.1.0",
 		"@types/material-ui": "^0.21.12",
 		"@types/react-dom": "^17.0.17",
-		"@types/styled-components": "^5.1.25",
+		"@types/styled-components": "^5.1.26",
 		"@typescript-eslint/eslint-plugin": "^4.22.0",
 		"@typescript-eslint/parser": "^4.22.0",
-		"babel-jest": "^26.6.3",
+		"babel-jest": "^29.1.0",
 		"babel-loader": "^8.2.2",
 		"babel-plugin-module-resolver": "^4.1.0",
 		"babel-plugin-react-remove-properties": "^0.3.0",
@@ -88,16 +88,16 @@
 		"eslint-plugin-react": "^7.23.2",
 		"eslint-plugin-storybook": "^0.5.3",
 		"husky": "^6.0.0",
-		"jest": "^26.6.3",
-		"jest-environment-jsdom-sixteen": "^1.0.3",
+		"jest": "^29.1.1",
+		"jest-environment-jsdom": "^29.1.1",
+		"jest-environment-jsdom-sixteen": "^2.0.0",
 		"jest-fetch-mock": "^3.0.3",
-		"jest-junit": "^12.0.0",
-		"jest-styled-components": "^7.0.4",
-		"jest-svg-transformer": "^1.0.0",
+		"jest-junit": "^14.0.1",
+		"jest-styled-components": "^7.1.1",
 		"lint-staged": "^10.5.4",
 		"npm-run-all": "^4.1.5",
 		"prettier": "^2.7.1",
-		"react-select-event": "^5.5.0",
+		"react-select-event": "^5.5.1",
 		"rxjs": "^7.2.0",
 		"standard-version": "^9.5.0",
 		"stylelint": "^13.13.1",
@@ -105,12 +105,12 @@
 		"stylelint-config-styled-components": "^0.1.1",
 		"stylelint-junit-formatter": "^0.2.2",
 		"stylelint-processor-styled-components": "^1.10.0",
-		"ts-jest": "^26.5.4",
+		"ts-jest": "^29.0.3",
 		"typescript": "^4.7.4"
 	},
 	"dependencies": {
-		"@amsterdam/asc-assets": "^0.36.0",
-		"@amsterdam/asc-ui": "^0.36.0",
+		"@amsterdam/asc-assets": "^0.37.0",
+		"@amsterdam/asc-ui": "^0.37.0",
 		"@mui/icons-material": "^5.10.3",
 		"@mui/material": "^5.10.5",
 		"@mui/styled-engine-sc": "^5.10.3",
