+++ src/form/AsyncSelect/__stubs__/options.ts
@@ -1,20 +1,6 @@
-export const options: { label: string; value: string }[] = [
-	'Archeologisch (voor)onderzoek',
-	'Archiefonderzoek',
-	'Bomeninventarisatie en –onderzoek',
-	'Bouwkundig onderzoek',
-	'Flora- en faunaonderzoek',
-	'(Geo)hydrologisch onderzoek',
-	'Geotechnisch onderzoek',
-	'Materiaalonderzoek',
-	'Milieukundig onderzoek',
-	'Onderzoek gevaarlijke stoffen',
-	'Toets Constructieve Veiligheid',
-	'Milieutechnisch onderzoek',
-].map((label) => ({
-	label,
-	value: label.toLowerCase().replace(/\W/g, ''),
-}));
+import { options as options_ } from '../../Select/__stubs__/options';
+
+export const options = options_;
 
 export function asyncOptions(value: string): Promise<{ label: string; value: string }[]> {
 	return new Promise((resolve) =>
