+++ src/form/CreatableSelect/CreatableSelect.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { ComponentProps } from 'react';
 import Creatable from 'react-select/creatable';
-import getAMSStyles, { StyledLabel } from './ReactSelectStyles';
+import { getSelectStyle, StyledLabel } from './ReactSelectStyles';
 
 type CreatableProps = ComponentProps<typeof Creatable>;
 export type Props = {
@@ -24,7 +24,7 @@ const CreatableSelect: React.FC<Props> = ({
 		<>
 			{label && <StyledLabel htmlFor={inputId} label={label} />}
 			<Creatable
-				styles={getAMSStyles({ error, ...props })}
+				styles={getSelectStyle({ error, ...props })}
 				value={value}
 				inputId={inputId}
 				openMenuOnFocus
