+++ src/form/index.ts
@@ -12,3 +12,4 @@ export { default as FileList } from './FileUpload/FileList/FileList';
 export { CustomFile } from './FileUpload/hooks';
 export { default as CreatableSelect } from './CreatableSelect/CreatableSelect';
 export { default as AsyncSelect } from './AsyncSelect/AsyncSelect';
+export { default as Select } from './Select/Select';
